modules/pam_timestamp: fix compilation warnings
Fix the following compilation warnings reported by gcc on ilp32 platforms:

  modules/pam_timestamp/hmacfile.c: In function ‘testvectors’:
  modules/pam_timestamp/hmacfile.c:121:44: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 2 has type ‘size_t’ {aka ‘unsigned int’} [-Wformat=]
    121 |      printf("Incorrect result for vector %lu\n", i + 1);
        |                                          ~~^     ~~~~~
        |                                            |       |
        |                                            |       size_t {aka unsigned int}
        |                                            long unsigned int
        |                                          %u
  modules/pam_timestamp/hmacfile.c:128:30: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 2 has type ‘size_t’ {aka ‘unsigned int’} [-Wformat=]
    128 |    printf("Error in vector %lu.\n", i + 1);
        |                            ~~^      ~~~~~
        |                              |        |
        |                              |        size_t {aka unsigned int}
        |                              long unsigned int
        |                            %u
  In function ‘strncpy’,
    inlined from ‘pam_sm_open_session’ at modules/pam_timestamp/pam_timestamp.c:584:4:
  /usr/include/bits/string_fortified.h:106:10: warning: ‘__builtin___strncpy_chk’ output may be truncated copying between 1 and 4095 bytes from a string of length 4095 [-Wstringop-truncation]

* modules/pam_timestamp/hmacfile.c (testvectors): Cast the argument
of type size_t to unsigned long before passing it to printf.
* modules/pam_timestamp/pam_timestamp.c (pam_sm_open_session): Use
memcpy instead of strncpy as the source is not NUL-terminated, add an
extra check to ensure that iterator stays inside bounds.

+++ modules/pam_timestamp/hmacfile.c
@@ -118,14 +118,16 @@ testvectors(void)
 				if (strncasecmp(hex,
 						vectors[i].hmac + 2 * j,
 						2) != 0) {
-					printf("Incorrect result for vector %lu\n", i + 1);
+					printf("Incorrect result for vector %lu\n",
+					       (unsigned long) i + 1);
 					exit(1);
 
 				}
 			}
 			free(hmac);
 		} else {
-			printf("Error in vector %lu.\n", i + 1);
+			printf("Error in vector %lu.\n",
+			       (unsigned long) i + 1);
 			exit(1);
 		}
 	}
